281: create new /look file structure
