--- src/helpers/constants.js
@@ -1,12 +1,0 @@
-import { findCachedTile } from './finders';
-
-
-let currGameIsCenterFlag;
-
-export function setupIsCenterFlag() {
-  currGameIsCenterFlag = findCachedTile(['YELLOW_FLAG', 'YELLOW_FLAG_TAKEN']) !== null;
-}
-
-export function isCenterFlag() {
-  return currGameIsCenterFlag;
-}
